filter record function is a single exported function
