+++ .eslintrc.json
@@ -18,6 +18,7 @@
     "function-paren-newline": 0,
     "no-plusplus": 0,
     "no-continue": 0,
-    "object-curly-newline": 0
+    "object-curly-newline": 0,
+    "operator-linebreak": 0
   }
 }
